requirements.txt
fastapi
uvicorn[standard]
gunicorn
streamlit
pandas
plotly
python-dotenv
pydantic
google-generativeai
google-adk
requests
sqlalchemy
